Rename default reaclibrary to logsfromnative package
